+++ megatron/io/__init__.py
@@ -1,3 +1,6 @@
 from . import generator
 from . import dataset
 from . import storage
+from .generator import *
+from .dataset import *
+from .storage import *